Finish bee-chat example
